Add color scheme for Search group
Gardener does not specify the color of Search group of terminal VIM,
it will use the default color "ctermbg=11 guibg=Yellow" (get from
':hi Search' in VIM). Note here that the default Search does not speicfy
foreground color, which means the foreground color are unchanged.
Consequently, the Yellow background makes the yellow texts not able to be seen.